Added new feature! Sharing existing instances with smart density
Gathers all existing instances, and copies them over to objects in "Instance" collection, if they don't have one of those mods yet. Also creates the vertex groups for them. The density is dynamically adjusted based on average edge length, so hopefully, density is not something that will have to be adjusted manually anymore.